requirements.txt
numpy
opencv-python
pillow
sounddevice
webrtcvad
SpeechRecognition
transformers
torch
nltk
gtts
deepface
python-pptx
pylatex
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
firebase-admin
pandas
openpyxl